+++ edstem/integration/storage/bh.cpp
@@ -71,7 +71,7 @@ void BinaryHeap::printBinaryHeap(int filter, int specifier, std::vector<std::str
     for (int i = 1; i < heapSize; i++) {
         printKeyFront(i, filter, specifier, types);
         // Iterate over both column names and elements in the column
-        for (int j = 1; j < heap[i]->data.second.first.size() - 1; j++) {
+        for (int j = 1; j < (int)heap[i]->data.second.first.size() - 1; j++) {
             
             if (filter == 0) {    
                 checkCol(i, j);
@@ -269,7 +269,7 @@ void BinaryHeap::printAllTypes() {
 void BinaryHeap::printAllTypesHelper(std::vector<std::string> &types, int &count) {
     for (int i = 1; i < heapSize; i++) {
         if (heap[i]->data.first != 0) {
-            for (int j = 0; j < heap[i]->data.second.first.size(); j++) {
+            for (int j = 0; j < (int)heap[i]->data.second.first.size(); j++) {
                 // make sure that j is 2 and that the type is not already in the vector
                 if (j == 2 && std::find(types.begin(), types.end(), heap[i]->data.second.first[j]) == types.end()) {
                     types.push_back(heap[i]->data.second.first[j]);
